requirements.txt
django==5.0.4
requests==2.31.0
djangorestframework==3.15.1
pillow==10.3.0
